tRecorderApi/requirements.txt
gunicorn==19.7
pydub
gevent
django==1.11
django_rest_framework
django-cors-headers
coreapi
urllib3
whitenoise
psycopg2
raven
celery
redis
django-celery-results
flower
drf-yasg
django-jenkins
coverage
